QR page converted to EJS template, rendering done in backend. All streaming service buttons added.
