+++ public/css/qr.css
@@ -7,8 +7,12 @@
     padding: 0;
 }
 
-a {
+a,
+a:hover,
+a:active,
+a:visited {
     text-decoration: none;
+    color: inherit;
 }
 
 body {
@@ -42,15 +46,11 @@ p {
     width: 400px;
 }
 
-.more {
-    font-size: .75em;
-    margin-top: 5px;
-}
-
-.newsflash {
-    font-size: 1.05em;
-    margin: 0 auto 15px;
-    width: 310px;
+.row {
+    display: flex;
+    flex-flow: row wrap;
+    margin: auto;
+    width: 99%;
 }
 
 .space-around {
@@ -65,26 +65,31 @@ p {
     justify-content: space-evenly;
 }
 
-.row {
-    display: flex;
-    flex-flow: row wrap;
-    margin: auto;
-    width: 99%;
-}
-
 /* Classes */
 .button {
     background-color: #000;
+    background-image: linear-gradient(rgba(0, 0, 0, 0.25), rgba(0, 0, 0, .45)), url("../images/black_glass.jpg");
+    background-position: center;
+    background-repeat: no-repeat;
+    background-size: 100%;
     border: 1px solid;
-    border-color: #d2d5db #bfc3cc #acb2bd;
+    border-color: #e5e7ea #d2d5db #bfc3cc;
     border-radius: 10px;
-    box-shadow: 0 1px 3px #3d5c73;
-    color: #fff;
-    height: 55px;
-    line-height: 55px;
-    margin: 15px auto;
-    width: 310px;
+    box-shadow: 0 0 60px 5px #253847;
+    color: #f7f7f7;
     cursor: pointer;
+    line-height: 53px;
+    margin: 15px auto;
+    text-shadow: 0 0 0 #000;
+    transition: all 0.25s ease-in-out;
+    width: 310px;
+}
+
+.button:hover,
+.button:active {
+    border-color: #eeeff2 #dbdee3 #c9ccd3 !important;
+    color: #fff !important;
+    text-shadow: 0 0 10px #555c6a !important;
 }
 
 .buttons {
@@ -99,7 +104,7 @@ p {
 }
 
 .hidden {
-    display: none;
+    display: none !important;
 }
 
 .icons {
@@ -112,28 +117,35 @@ p {
 .listen-btn {
     display: block;
     margin: 5px 5px;
-    width: 150px;
+    max-width: 150px;
 }
 
-.modal {
-    display: flex;
-    width: 300px;
-    height: 300px;
-    background-color: #fff;
-    position: fixed;
-    margin: 20px auto;
-    left: -50%;
-    right: -50%;
-    /* text-align: center; */
-
+.more {
+    cursor: pointer;
+    display: block;
+    font-size: .75em;
+    margin-top: 5px;
+    text-shadow: 1px 1px 5px #000;
 }
 
 .pentup-logo {
+    position: relative;
     margin: 10px;
-    width: 200px;
+    width: 225px;
+    z-index: -1;
 }
 
 .social-icon {
     height: 40px;
     cursor: pointer;
+}
+
+.news span {
+    position: relative;
+    font-size: 1.05em;
+    margin: auto;
+    text-shadow: 0 0 0 #000;
+    width: 310px;
+    overflow: hidden;
+    top: -10px;
 }
