+++ views/partials/head.ejs
@@ -1,0 +1,22 @@
+<head>
+    <meta charset="UTF-8">
+    <meta http-equiv="X-UA-Compatible" content="IE=edge">
+    <meta name="viewport" content="width=device-width, initial-scale=1.0">
+    <link rel="stylesheet" href="/css/qr.css">
+    <script src="/javascript/<%- script %>" defer></script>
+    <title><%-title%></title>
+    <link rel="apple-touch-icon" sizes="180x180" href="/images/favicons/apple-touch-icon.png">
+    <link rel="icon" type="image/png" sizes="32x32" href="/images/favicons/favicon-32x32.png">
+    <link rel="icon" type="image/png" sizes="16x16" href="/images/favicons/favicon-16x16.png">
+    <link rel="manifest" href="/images/favicons/site.webmanifest">
+    <link rel="mask-icon" href="/images/favicons/safari-pinned-tab.svg" color="#5bbad5">
+    <link rel="shortcut icon" href="/images/favicons/favicon.ico">
+    <meta name="msapplication-TileColor" content="#c191e4">
+    <meta name="msapplication-config" content="/images/favicons/browserconfig.xml">
+    <meta name="theme-color" content="#ffffff">
+    <meta name="author" content="Scruffy Minds Web Design">
+    <meta name="description" content="Pent Up! is a band that is here to drink beer and play punk rock... and they're all out of beer.">
+    <meta name="keywords" content="Band, Alternative, Rock, Punk, Punk Rock, Emo, Indie, Music, Seattle, Washington, Pacific Northwest">
+    <meta name="viewport" content="width=device-width, initial-scale=1.0">
+    <%- include('gtag'); -%>
+</head>
