+++ views/qr.ejs
@@ -1,37 +1,8 @@
 <!DOCTYPE html>
 <html lang="en">
-
-<head>
-    <meta charset="UTF-8">
-    <meta http-equiv="X-UA-Compatible" content="IE=edge">
-    <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <link rel="stylesheet" href="/css/qr.css">
-    <script src="/javascript/qr.js" defer></script>
-    <title><%-title%></title>
-    <link rel="apple-touch-icon" sizes="180x180" href="/images/favicons/apple-touch-icon.png">
-    <link rel="icon" type="image/png" sizes="32x32" href="/images/favicons/favicon-32x32.png">
-    <link rel="icon" type="image/png" sizes="16x16" href="/images/favicons/favicon-16x16.png">
-    <link rel="manifest" href="/images/favicons/site.webmanifest">
-    <link rel="mask-icon" href="/images/favicons/safari-pinned-tab.svg" color="#5bbad5">
-    <link rel="shortcut icon" href="/images/favicons/favicon.ico">
-    <meta name="msapplication-TileColor" content="#c191e4">
-    <meta name="msapplication-config" content="/images/favicons/browserconfig.xml">
-    <meta name="theme-color" content="#ffffff">
-    <!-- Global site tag (gtag.js) - Google Analytics -->
-    <script async src="https://www.googletagmanager.com/gtag/js?id=G-7QT7P357RK"></script>
-    <script>
-        window.dataLayer = window.dataLayer || [];
-
-        function gtag() {
-            dataLayer.push(arguments);
-        }
-        gtag('js', new Date());
-
-        gtag('config', 'G-7QT7P357RK');
-    </script>
-</head>
-
+<%- include('partials/head'); -%>
 <body>
+
     <div class="container">
 
         <div class="row space-around">
@@ -39,18 +10,34 @@
                 <img class="pentup-logo" src="/images/pent_up_logo_small.png" alt="Pent Up!">
             </a>
         </div>
-        <!-- <div class="row center modal">coming soon</div> -->
-        <div class="row center">
-            <span class="newsflash"></span>
+
+        <div class="news row center">
+            <span><%= news %></span>
         </div>
+
         <div id="streaming" class="row center">
-            <!-- <span class="more">Tap for more streaming services</span> -->
+            <%const streamers = sites.filter(x => x.type === 'streaming');  streamers.forEach((v, i, a) => {const primary = (v.id === 'applemusic' || v.id === 'spotify') ? ' primary' : ' hidden';%>
+            <%-`<a href="${v.url}" target="${v.target}"><img id="${v.id}" class="listen-btn button${primary}" src="/images/buttons/${v.id}_btn.png" alt="Listen on ${v.name}"></a>`%><%});%>
+            <span class="more row center">Tap here for more streaming services</span>
+        
         </div>
+
         <div id="buttons" class="buttons">
+            <%const websites = sites.filter(x => x.type === 'website'); websites.forEach((v, i, a) => {%>
+            <%- `<a href="${v.url}" target="${v.target}"><div class="button" id="${v.id}"><span>${v.name}</span></div></a>`%><%})%>
+        
         </div>
+
         <div id="socials" class="row space-around icons">
+        <%const socials = sites.filter(x => x.type === 'social'); socials.forEach((v, i, a) => {%>
+            <%- `<a href="${v.url}" target="${v.target}"><img id="${v.id}" class="social-icon" src="/images/icons/icon-${v.id}.png" alt="${v.name}"></a>`%><%})%>
+        
         </div>
-        <div class="row center"><span class="copyright"></span></div>
+        <% const thisYear = new Date().getFullYear(); %>
+        <div class="row center">
+            <span class="copyright">&copy;<%- thisYear %>, Pent Up! All rights reserved.</span>
+        </div>
+
     </div>
 </body>
 </html>
