+++ views/qr_old.ejs
@@ -1,0 +1,58 @@
+<!DOCTYPE html>
+<html lang="en">
+
+<head>
+    <meta charset="UTF-8">
+    <meta http-equiv="X-UA-Compatible" content="IE=edge">
+    <meta name="viewport" content="width=device-width, initial-scale=1.0">
+    <link rel="stylesheet" href="/css/qr.css">
+    <script src="/javascript/qr.js" defer></script>
+    <title><%-title%></title>
+    <link rel="apple-touch-icon" sizes="180x180" href="/images/favicons/apple-touch-icon.png">
+    <link rel="icon" type="image/png" sizes="32x32" href="/images/favicons/favicon-32x32.png">
+    <link rel="icon" type="image/png" sizes="16x16" href="/images/favicons/favicon-16x16.png">
+    <link rel="manifest" href="/images/favicons/site.webmanifest">
+    <link rel="mask-icon" href="/images/favicons/safari-pinned-tab.svg" color="#5bbad5">
+    <link rel="shortcut icon" href="/images/favicons/favicon.ico">
+    <meta name="msapplication-TileColor" content="#c191e4">
+    <meta name="msapplication-config" content="/images/favicons/browserconfig.xml">
+    <meta name="theme-color" content="#ffffff">
+    <meta name="author" content="Scruffy Minds Web Design">
+    <meta name="description" content="Pent Up! is a band that is here to drink beer and play punk rock... and they're all out of beer.">
+    <meta name="keywords" content="Band, Alternative, Rock, Punk, Punk Rock, Emo, Indie, Music, Seattle, Washington, Pacific Northwest">
+    <meta name="viewport" content="width=device-width, initial-scale=1.0">
+    <!-- Global site tag (gtag.js) - Google Analytics -->
+    <script async src="https://www.googletagmanager.com/gtag/js?id=G-7QT7P357RK"></script>
+    <script>
+        window.dataLayer = window.dataLayer || [];
+
+        function gtag() {
+            dataLayer.push(arguments);
+        }
+        gtag('js', new Date());
+
+        gtag('config', 'G-7QT7P357RK');
+    </script>
+</head>
+
+<body>
+    <div class="container">
+
+        <div class="row space-around">
+            <a href="http://www.pentup.band" target="_blank">
+                <img class="pentup-logo" src="/images/pent_up_logo_small.png" alt="Pent Up!">
+            </a>
+        </div>
+        <div class="news row center">
+            <span></span>
+        </div>
+        <div id="streaming" class="row center">
+        </div>
+        <div id="buttons" class="buttons">
+        </div>
+        <div id="socials" class="row space-around icons">
+        </div>
+        <div class="row center"><span class="copyright"></span></div>
+    </div>
+</body>
+</html>
